requirements.txt
certifi==2020.12.5
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
olefile==0.46
pillow==8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
torchaudio
torchvision
tornado==6.1
typing_extensions==3.7.4.3
six==1.15.0
